Add: When generating source code "source -f <graphml file > -t <template file> the data field Data/Description in yEd will be added as a comment. See src/test/resources/templates/junit.template which uses the new {DESCRIPTION}
Did cleanup and formatting of code

+++ src/test/java/org/graphwalker/conditions/AlternativeConditionTest.java
@@ -37,11 +37,11 @@ public class AlternativeConditionTest extends TestCase {
 	public void testFulfillment() {
 		AlternativeCondition condition = new AlternativeCondition();
 		condition.add(new NeverCondition());
-		assertEquals((double) 0, condition.getFulfilment(), 0.01);
+		assertEquals(0, condition.getFulfilment(), 0.01);
 		condition.add(new NeverCondition());
-		assertEquals((double) 0, condition.getFulfilment(), 0.01);
+		assertEquals(0, condition.getFulfilment(), 0.01);
 		condition.add(new NeverCondition());
-		assertEquals((double) 0, condition.getFulfilment(), 0.01);
+		assertEquals(0, condition.getFulfilment(), 0.01);
 	}
 
 	public void testIsFulfilled() {
